+++ src/loaders/JsonLoader.js
@@ -1,6 +1,6 @@
 import Ajv from 'ajv';
-import AbstractLoader from './AbstractLoader';
-import { LoaderFetchError, LoaderValidationError, AbstractLoaderError } from './errors/index';
+import AbstractTwoStepLoader from './AbstractTwoStepLoader';
+import { LoaderValidationError, AbstractLoaderError } from './errors/index';
 import LoaderResult from './LoaderResult';
 
 import cellsSchema from '../schemas/cells.schema.json';
@@ -17,27 +17,12 @@ const typeToSchema = {
   'cell-sets': cellSetsSchema,
 };
 
-export default class JsonLoader extends AbstractLoader {
-  constructor(params) {
-    super(params);
+export default class JsonLoader extends AbstractTwoStepLoader {
+  constructor(dataSource, params) {
+    super(dataSource, params);
 
     const { type } = params;
     this.schema = typeToSchema[type];
-  }
-
-  loadJson() {
-    const {
-      url, requestInit, type, fileType,
-    } = this;
-    return fetch(url, requestInit)
-      .then((response) => {
-        if (response.ok) {
-          return response.json();
-        }
-        return Promise.reject(
-          new LoaderFetchError(type, fileType, url, response.headers),
-        );
-      });
   }
 
   load() {
@@ -47,7 +32,7 @@ export default class JsonLoader extends AbstractLoader {
     if (this.data) {
       return this.data;
     }
-    this.data = this.loadJson()
+    this.data = this.dataSource.data
       .then((data) => {
         if (data instanceof AbstractLoaderError) {
           return Promise.reject(data);