+++ src/schemas/schema-schema.sh
@@ -20,3 +20,15 @@ if [[ $FAILURES ]]; then
   die "Check that 'additionalProperties' & 'required' & 'properties' are provided:
 $FAILURES"
 fi
+
+# Check that existing schemas have not changed.
+for SCHEMA in $SCHEMA_DIR/config-*.schema.json; do
+  SCHEMA_NAME=`basename $SCHEMA`
+  SCHEMA_RESPONSE=`curl -s https://raw.githubusercontent.com/vitessce/vitessce/master/src/schemas/$SCHEMA_NAME`
+  if [ "$SCHEMA_RESPONSE" == "404: Not Found" ]; then
+    echo "Schema $SCHEMA is not in production yet."
+  else
+    diff --ignore-all-space $SCHEMA <(echo "$SCHEMA_RESPONSE") \
+      || die "$SCHEMA has changed.  Please do not change existing schema; instead, make a new schema."
+  fi
+done