+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 ### Changed
 - Updated `GeoJSON` loader to support the Baysor custom format.
 
+## [1.1.13](https://www.npmjs.com/package/vitessce/v/1.1.13) - 2021-08-10
+
+### Added
+- `disableChannelsIfRgbDetected` prop for turning of channel controllers when rgb is detected.
+
+### Changed
+- Introduce two-step data loaders for AnnData "files".
+- Update README to have more info on using view configs via url parameters.
+- Add a check for schema changes - schemas may not be updated once published, only new ones added.
+- Make two-step data loaders universal for all file types.
+- Upgrade `higlass-zarr-datafetchers` to 0.2.1 to prevent the latest Zarr.js from making failed HEAD requests.
+- Replace "hubmapconsortium/vitessce" with "vitessce/vitessce"
+
 ## [1.1.12](https://www.npmjs.com/package/vitessce/v/1.1.12) - 2021-07-20
 
 ### Added